improve image opening from file
Now the function assumes that the argument given is the absolute
path and only if that doesn't exist it looks into the 'images'
folder. To find out the location of the images folder, the CMake
creates a file called 'src/location.hpp' which gives the absolute
location of the source folder. This is a workaround to the problem
with relative paths.